+++ R/forest.cumul.rma.r
@@ -3,7 +3,7 @@ annotate=TRUE,                                                header=FALSE,
 xlim, alim, olim, ylim, at, steps=5, level=x$level, refline=0, digits=2L, width,
 xlab,             ilab, ilab.xpos, ilab.pos,
 transf, atransf, targs, rows,
-efac=1, pch, psize,                           col,
+efac=1, pch, psize,                           col,         shade, colshade,
 lty, fonts, cex, cex.lab, cex.axis, ...) {
 
    #########################################################################
@@ -68,6 +68,15 @@ lty, fonts, cex, cex.lab, cex.axis, ...) {
    } else {
       psize <- .getx("psize", mf=mf, data=x$data)
    }
+
+   if (missing(shade)) {
+      shade <- NULL
+   } else {
+      shade <- .getx("shade", mf=mf, data=x$data)
+   }
+
+   if (missing(colshade))
+      colshade <- "lightgray"
 
    if (missing(cex))
       cex <- NULL
@@ -172,14 +181,28 @@ lty, fonts, cex, cex.lab, cex.axis, ...) {
       top <- ddd$top
    }
 
-   lplot     <- function(..., textpos, clim, rowadj, annosym, top) plot(...)
-   labline   <- function(..., textpos, clim, rowadj, annosym, top) abline(...)
-   lsegments <- function(..., textpos, clim, rowadj, annosym, top) segments(...)
-   laxis     <- function(..., textpos, clim, rowadj, annosym, top) axis(...)
-   lmtext    <- function(..., textpos, clim, rowadj, annosym, top) mtext(...)
-   lpolygon  <- function(..., textpos, clim, rowadj, annosym, top) polygon(...)
-   ltext     <- function(..., textpos, clim, rowadj, annosym, top) text(...)
-   lpoints   <- function(..., textpos, clim, rowadj, annosym, top) points(...)
+   if (is.null(ddd$xlabadj)) {
+      xlabadj <- c(NA,NA)
+   } else {
+      xlabadj <- ddd$xlabadj
+      if (length(xlabadj) == 1L)
+         xlabadj <- c(xlabadj, 1-xlabadj)
+   }
+
+   if (is.null(ddd$xlabfont)) {
+      xlabfont <- 1
+   } else {
+      xlabfont <- ddd$xlabfont
+   }
+
+   lplot     <- function(..., textpos, clim, rowadj, annosym, top, xlabadj, xlabfont) plot(...)
+   labline   <- function(..., textpos, clim, rowadj, annosym, top, xlabadj, xlabfont) abline(...)
+   lsegments <- function(..., textpos, clim, rowadj, annosym, top, xlabadj, xlabfont) segments(...)
+   laxis     <- function(..., textpos, clim, rowadj, annosym, top, xlabadj, xlabfont) axis(...)
+   lmtext    <- function(..., textpos, clim, rowadj, annosym, top, xlabadj, xlabfont) mtext(...)
+   lpolygon  <- function(..., textpos, clim, rowadj, annosym, top, xlabadj, xlabfont) polygon(...)
+   ltext     <- function(..., textpos, clim, rowadj, annosym, top, xlabadj, xlabfont) text(...)
+   lpoints   <- function(..., textpos, clim, rowadj, annosym, top, xlabadj, xlabfont) points(...)
 
    #########################################################################
 
@@ -236,6 +259,28 @@ lty, fonts, cex, cex.lab, cex.axis, ...) {
    if (length(col) != k)
       stop(mstyle$stop(paste0("Length of the 'col' argument (", length(col), ") does not correspond to the number of outcomes (", k, ").")))
 
+   shade.type <- "none"
+
+   if (is.character(shade)) {
+
+      shade.type <- "character"
+      shade <- shade[1]
+
+      if (!is.element(shade, c("zebra", "zebra1", "zebra2", "all")))
+         stop(mstyle$stop("Unknown option specified for 'shade' argument."))
+
+   }
+
+   if (is.logical(shade)) {
+
+      shade.type <- "logical"
+      shade <- .chksubset(shade, k, stoponk0=FALSE)
+
+   }
+
+   if (is.numeric(shade))
+      shade.type <- "numeric"
+
    ### set rows value
 
    if (missing(rows)) {
@@ -260,6 +305,8 @@ lty, fonts, cex, cex.lab, cex.axis, ...) {
    psize   <- psize[k:1]                          # if NULL, remains NULL
    col     <- col[k:1]
    rows    <- rows[k:1]
+   if (shade.type == "logical")
+      shade <- shade[k:1]
 
    ### check for NAs in yi/vi and act accordingly
 
@@ -279,6 +326,8 @@ lty, fonts, cex, cex.lab, cex.axis, ...) {
          pch     <- pch[not.na]
          psize   <- psize[not.na]               # if NULL, remains NULL
          col     <- col[not.na]
+         if (shade.type == "logical")
+            shade <- shade[not.na]
 
          rows.new <- rows                       # rearrange rows due to NAs being omitted from plot
          rows.na  <- rows[!not.na]              # shift higher rows down according to number of NAs omitted
@@ -468,6 +517,29 @@ lty, fonts, cex, cex.lab, cex.axis, ...) {
 
    lplot(NA, NA, xlim=xlim, ylim=ylim, xlab="", ylab="", yaxt="n", xaxt="n", xaxs="i", bty="n", ...)
 
+   if (shade.type == "character") {
+      if (shade == "zebra" || shade == "zebra1")
+         tmp <- rep_len(c(TRUE,FALSE), k)
+      if (shade == "zebra2")
+         tmp <- rep_len(c(FALSE,TRUE), k)
+      if (shade == "all")
+         tmp <- rep_len(TRUE, k)
+      shade <- tmp
+   }
+
+   if (shade.type %in% c("character","logical")) {
+      for (i in seq_len(k)) {
+         if (shade[i])
+            rect(xlim[1], rows[i]-0.5, xlim[2], rows[i]+0.5, border=colshade, col=colshade)
+      }
+   }
+
+   if (shade.type == "numeric") {
+      for (i in seq_along(shade)) {
+         rect(xlim[1], shade[i]-0.5, xlim[2], shade[i]+0.5, border=colshade, col=colshade)
+      }
+   }
+
    ### horizontal title line
 
    labline(h=ylim[2]-(top-1), lty=lty[2], ...)
@@ -512,7 +584,20 @@ lty, fonts, cex, cex.lab, cex.axis, ...) {
    if (missing(xlab))
       xlab <- .setlab(measure, transf.char, atransf.char, gentype=2)
 
-   lmtext(xlab, side=1, at=min(at) + (max(at)-min(at))/2, line=par("mgp")[1]-0.5, cex=cex.lab, ...)
+   if (!is.element(length(xlab), 1:3))
+      stop(mstyle$stop("Argument 'xlab' argument must be of length 1, 2, or 3."))
+
+   if (length(xlab) == 1L)
+      lmtext(xlab, side=1, at=min(at) + (max(at)-min(at))/2, line=par("mgp")[1]-0.5, cex=cex.lab, col="black", font=xlabfont[1], ...)
+   if (length(xlab) == 2L) {
+      lmtext(xlab[1], side=1, at=min(at), line=par("mgp")[1]-0.5, cex=cex.lab, col="black", adj=xlabadj[1], font=xlabfont[1], ...)
+      lmtext(xlab[2], side=1, at=max(at), line=par("mgp")[1]-0.5, cex=cex.lab, col="black", adj=xlabadj[2], font=xlabfont[1], ...)
+   }
+   if (length(xlab) == 3L) {
+      lmtext(xlab[1], side=1, at=min(at), line=par("mgp")[1]-0.5, cex=cex.lab, col="black", adj=xlabadj[1], font=xlabfont[1], ...)
+      lmtext(xlab[2], side=1, at=min(at) + (max(at)-min(at))/2, line=par("mgp")[1]-0.5, cex=cex.lab, col="black", font=xlabfont[2], ...)
+      lmtext(xlab[3], side=1, at=max(at), line=par("mgp")[1]-0.5, cex=cex.lab, col="black", adj=xlabadj[2], font=xlabfont[1], ...)
+   }
 
    ### add CI ends (either | or <> if outside of axis limits)