+++ NEWS.md
@@ -1,6 +1,12 @@
-# metafor 4.1-2 (2023-04-04)
+# metafor 4.1-3 (2023-04-07)
 
-- version bump for development version
+- also allow two or three values for `xlab` in the various `forest()` functions (for adding labels at the ends of the x-axis limits)
+
+- added `shade` and `colshade` arguments to the various `forest()` functions
+
+- added `link` argument to `rma.glmm()`
+
+- small bug fixes
 
 # metafor 4.0-0 (2023-03-19)